# Some alignment optimizations