<!-- Feel free to remove check-list items aren't relevant to your change -->

- [x] Benchmark added
- [x] User visible changes (including notable bug fixes) are documented in `whats-new.rst`

May fix some performance regressions, e.g., see https://github.com/pydata/xarray/issues/7376#issuecomment-1352989233.

@ravwojdyla with this PR `ds.assign(foo=~ds["d3"])` in your example should be much faster (on par with version 2022.3.0).